add adpator && fix eplb_updator
